create order

+++ src/dao/UserDao.java
@@ -96,7 +96,7 @@ public class UserDao {
         return User.resultSetToUser(resultSet);
     }
 
-    public void updateUserById(Long id, String address) throws SQLException {
+    public void updateUserById(int id, String address) throws SQLException {
         if(!existsById(id)) {
             throw new UserNotExistsException("There is no user with such id");
         }
@@ -105,7 +105,7 @@ public class UserDao {
                 .prepareStatement("UPDATE users SET address = ? WHERE id = ?");
 
         preparedStatement.setString(1, address);
-        preparedStatement.setLong(2, id);
+        preparedStatement.setInt(2, id);
 
         preparedStatement.executeUpdate();
     }
@@ -119,10 +119,10 @@ public class UserDao {
         return resultSet.next();
     }
 
-    public boolean existsById(Long id) throws SQLException{
+    public boolean existsById(int id) throws SQLException{
         PreparedStatement existsById = connection.
                 prepareStatement("SELECT 1 FROM users WHERE id = ?");
-        existsById.setLong(1, id);
+        existsById.setInt(1, id);
         ResultSet resultSet = existsById.executeQuery();
 
         return resultSet.next();
